Allow using extensions from build output
