+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup/BuildOutputSharedLibraryInitializer.cs
@@ -5,17 +5,11 @@ using Microsoft.Diagnostics.Tools.Monitor;
 using Microsoft.Diagnostics.Tools.Monitor.LibrarySharing;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Logging;
-using System.IO;
-using System.Reflection;
 
 namespace Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup
 {
     internal sealed class BuildOutputSharedLibraryInitializer : ISharedLibraryInitializer
     {
-        // This is the binary output directory when built from the dotnet-monitor repo: <repoRoot>/artifacts/bin
-        private static readonly string SharedLibrarySourcePath =
-            Path.GetFullPath(Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "..", "..", ".."));
-
         private readonly ILogger<BuildOutputSharedLibraryInitializer> _logger;
 
         public BuildOutputSharedLibraryInitializer(ILogger<BuildOutputSharedLibraryInitializer> logger)
@@ -25,7 +19,7 @@ namespace Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup
 
         public IFileProviderFactory Initialize()
         {
-            _logger.SharedLibraryPath(SharedLibrarySourcePath);
+            _logger.SharedLibraryPath(BuildOutput.RootPath);
 
             return new Factory();
         }
@@ -34,12 +28,12 @@ namespace Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup
         {
             public IFileProvider CreateManaged(string targetFramework)
             {
-                return BuildOutputManagedFileProvider.Create(targetFramework, SharedLibrarySourcePath);
+                return BuildOutputManagedFileProvider.Create(targetFramework, BuildOutput.RootPath);
             }
 
             public IFileProvider CreateNative(string runtimeIdentifier)
             {
-                return BuildOutputNativeFileProvider.Create(runtimeIdentifier, SharedLibrarySourcePath);
+                return BuildOutputNativeFileProvider.Create(runtimeIdentifier, BuildOutput.RootPath);
             }
         }
     }
