+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup/Microsoft.Diagnostics.Monitoring.Tool.TestHostingStartup.csproj
@@ -5,6 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\Microsoft.Diagnostics.Monitoring.TestCommon\TargetFrameworkMoniker.cs" Link="TargetFrameworkMoniker.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
